refactor(script): split effect buildins into buildin_effect.c
Issue #61 tenth step: move misceffect/soundeffect/musiceffect helpers
out of script.c.
