+++ vcproj/map-server.vcxproj.filters
@@ -133,6 +133,9 @@
     <ClCompile Include="..\src\map\buildin_quest.c">
       <Filter>map_sql</Filter>
     </ClCompile>
+    <ClCompile Include="..\src\map\buildin_effect.c">
+      <Filter>map_sql</Filter>
+    </ClCompile>
     <ClCompile Include="..\src\map\skill.c">
       <Filter>map_sql</Filter>
     </ClCompile>
